moved CSS files to CSS folder
